pyproject.toml
[tool.black]
line-length = 100

[tool.isort]
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.flake8]
ignore = ["E402","W503","F811", "F401"]
max-line-length = 100
exclude = "_ignore"

[tool.mypy]
exclude = "_ignore"

[tool.poetry]
name = "age of gold"
version = "0.2.0"
description = "Age of Gold is a fun trading game. The first single player game that you play with multiple people"
authors = ["Grabot"]
license = "GPL-3.0-or-later"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.10"
fastapi = "0.116.1"
uvicorn = "0.35.0"
sqlmodel = "0.0.24"
alembic = "1.16.4"
python-socketio = "5.13.0"
websockets = "15.0.1"
asyncio-redis = "0.16.0"
redis = "6.4.0"
fastapi-pagination = "0.14.0"
celery = "5.5.3"
pydantic-settings = "2.10.1"
asyncpg = "0.30.0"
psycopg2-binary = "2.9.10"
APScheduler = "3.11.0"
sqlalchemy = "2.0.43"
argon2-cffi = "^25.1.0"
pyjwt = "^2.10.1"
pillow = "^11.3.0"
requests = "^2.32.5"
types-requests = "^2.32.4.20250809"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.9"
pre-commit = "^4.3.0"
mypy = "^1.17.1"
flake8 = "^7.3.0"
python-socketio-stubs = "^5.13.0.12.20250809"
celery-stubs = "^0.1.3"
sqlalchemy-stubs = "^0.4"
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
pytest-mock = "^3.14.1"
isort = "^6.0.1"
aiosqlite = "^0.21.0"
httpx = "^0.28.1"
podman-compose = "^1.5.0"
Flake8-pyproject = "1.2.3"
bump-pydantic = "^0.8.0"
types-requests = "^2.32.4.20250809"
cryptography = "^45.0.6"
pytest-cov = "^6.2.1"
sqlalchemy2-stubs = "^0.0.2a38"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
